Rewrote IF implementation
Previously IF skipped the evaluation of an expression on a false value.
This stopped working inside functions.
IF relied on the next expression being the next one evaluated, but when a function is called,
only a single expression gets run, meaning IF's branch never got run.

Now IF works more like a user defined function.
It evaluates its branch itselfs (on a true value), and then moves the index to skip it.

Here's an example of how the IF function used to cause issues:

\#Does print
FN "MAIN" ""
	; IF true ECHO "PRINT 1"
	null

MAIN

\#Does not
FN "MAIN" ""
	IF true ECHO "PRINT 2"

MAIN

\#Does not print the second
FN "MAIN" ""
	; IF true ECHO "PRINT 3"
	ECHO "PRINT 4"

MAIN